feat: complete Keystone V1 pilot launch package
- Refactored core entry screens (LogJob, AddCustomer, AddNote) into multi-step tactical wizards.
- Implemented Admin Correction system with in-app approval dashboard and specialized RLS.
- Enhanced UI with 'Military Grade' polish: Haptic feedback, Monospace typography, and Field Hints.
- Strengthened Offline-First reliability with Hive corruption auto-recovery and global sync badges.
- Resolved profile URL 404 issue and standardized identity mapping.
- Completed automated test suite with 80+ passing tests (Admin logic, Wizard flows, Sync integration).
- Fully updated V1 documentation: schema, dev log, current state, and admin runbook.

+++ lib/features/job_logging/domain/entities/job_entity.dart
@@ -10,7 +10,7 @@ class JobEntity {
   final double? latitude;
   final double? longitude;
   final String? notes;
-  final double? amountCharged;
+  final int? amountCharged;
   final bool followUpSent;
   final DateTime? followUpSentAt;
   final SyncStatus syncStatus;
@@ -56,7 +56,7 @@ class JobEntity {
     double? latitude,
     double? longitude,
     String? notes,
-    double? amountCharged,
+    int? amountCharged,
     bool? followUpSent,
     DateTime? followUpSentAt,
     SyncStatus? syncStatus,
